# docx/product-factory/07_dev_contract/final/openapi.yaml
openapi: 3.0.3
info:
  title: Nexa API
  version: 1.0.0
  description: |
    Nexa — AI API 网关 SaaS（基于 new-api 二次开发，Go + Gin + GORM）的机读 API 契约。
    本文件由人读版权威契约 `API-ENDPOINTS.md`（§1-§16）翻译而来，供前端 mock / 生成 client 用。

    产品铁律（贯穿所有出参 DTO）：
    1. 客户视图 DTO（*UserView/*PublicView）绝不含：成本（quota_cost）、利润（quota_profit）、
       上游真实模型 B（actual_upstream_model / upstream_name / upstream_model）、供应商渠道维
       （channel_id / channel_name）、cost_ratio。
    2. 管理视图 DTO（*AdminView）给全链 C→A→B + 协议 + 售价/成本/利润 + channel_id。
    3. 凡涉及 key/secret 的字段：列表脱敏（MaskTokenKey），取明文走 CriticalRateLimit + DisableCache + 二次验证。

    通用响应包络（new-api 惯例）：{ "success": bool, "message": string, "data": <payload> }。
    分页响应 data 形如 { "items": [...], "total": int, "page": int, "page_size": int }。

    Relay 对外 schema 铁律：所有 /v1/*、/v1beta/* 入口的对外 API schema = 标准 OpenAI / Anthropic / Gemini
    原生格式；内部 C→A→B 两层映射、选渠、协议头尾转换、双价记账对客户完全不感知。
servers:
  - url: https://nexa.ai
    description: Nexa 生产环境
security: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: sk-nexa-...
      description: |
        API Key（sk-nexa 前缀）鉴权，用于 /v1/* 与 /v1beta/* 的 relay 中转端点（TokenAuth）。
        Header: `Authorization: Bearer sk-nexa-...`。Claude 原生另支持 `x-api-key` + `anthropic-version`，
        Gemini 原生另支持 `x-goog-api-key` 或 query key。
    tokenReadAuth:
      type: http
      scheme: bearer
      bearerFormat: sk-nexa-...
      description: |
        TokenAuthReadOnly：Bearer sk-key 只读鉴权，用于按 key 查用量/日志（/api/usage/token、/api/log/token）。
        附加 CriticalRateLimit + DisableCache。
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: |
        UserAuth（会话/cookie 或 access token）。用于 /api/user/self/* 等 self-scope 端点，
        按 user_id 强制过滤越权。普通用户角色（common+）。部分高危动作要求 dashboard 会话（拒绝 access token）。
    adminAuth:
      type: apiKey
      in: cookie
      name: session
      description: |
        AdminAuth（admin+ 角色）。基于会话鉴权，要求 Role >= admin。用于用户管理、渠道/模型/计费配置、
        管理端日志、部署管理等。越权路由返回 403（F-5031）。文档层用同一会话凭据 + 角色级别区分。
    rootAuth:
      type: apiKey
      in: cookie
      name: session
      description: |
        RootAuth（root 角色）。基于会话鉴权，要求 Role == root。用于全站 Option 配置、性能监控运维、
        支付合规确认、自定义 OAuth provider 配置等。部分端点额外要求 dashboard 会话。
  schemas:
...
    ApiResponse:
      type: object
      description: 通用响应包络（new-api 惯例）
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          nullable: true
          description: 业务载荷，按端点不同而不同
      required:
        - success
    SuccessResponse:
      type: object
      description: 仅 success 的成功响应（多数状态变更/删除端点）
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          nullable: true
      required:
        - success
    Pagination:
      type: object
      description: 分页元数据（data 内）
      properties:
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
    ErrorResponse:
      type: object
      description: 业务错误响应（success=false）
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        data:
          nullable: true
      required:
        - success
        - message

    # ============ 账号与身份 DTO（模块一） ============
...
    UserAdminView:
      type: object
      description: |
        用户管理视图（AdminAuth）。含 remark / inviter_id（区别于 UserView）。
      properties:
        id:
          type: integer
        username:
          type: string
        display_name:
          type: string
        role:
          type: integer
        status:
          type: integer
        email:
          type: string
        group:
          type: string
        quota:
          type: integer
        used_quota:
          type: integer
        request_count:
          type: integer
        aff_code:
          type: string
        aff_count:
          type: integer
        inviter_id:
          type: integer
        remark:
          type: string
        github_id:
          type: string
        discord_id:
          type: string
        wechat_id:
          type: string
        telegram_id:
          type: string
        last_login_at:
          type: integer
          format: int64
        created_at:
          type: integer
          format: int64
        setting:
          type: object
          additionalProperties: true
          description: 单用户详情时含 setting
    AdminBalanceAdjustRequest:
      type: object
      description: 管理员充值/扣费请求（amount 以 USD 计）
      properties:
        amount:
          type: number
          description: USD 金额（> 0）
        remark:
          type: string
      required: [amount]
    BalanceTransactionView:
      type: object
      description: 账变流水视图（amount/balance_after 以 USD 计）
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: [ADMIN_CREDIT, ADMIN_DEBIT, REDEEM, TOPUP]
        amount:
          type: number
          description: 变动额（USD，带正负）
        balance_after:
          type: number
          description: 变动后余额（USD）
        operator_id:
          type: integer
          format: int64
        remark:
          type: string
        created_time:
          type: integer
          format: int64
...
  responses:
...
    ForbiddenError:
      description: 越权/权限不足（403）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequestError:
      description: 参数错误 / 业务校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  parameters:
    PageParam:
      name: p
      in: query
      description: 页码
      schema:
        type: integer
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
...
paths:
...
  /api/user/:
    get:
      tags: [account]
      summary: 用户列表（分页）
      x-function-id: F-1008
      security:
        - adminAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: 用户列表（AdminView）
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/Pagination'
                          - type: object
                            properties:
                              items:
                                type: array
                                items:
                                  $ref: '#/components/schemas/UserAdminView'
    post:
      tags: [account]
      summary: 创建用户
      x-function-id: F-1009
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                display_name:
                  type: string
                role:
                  type: integer
                  description: 不可高于自身角色
              required: [username, password]
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
    put:
      tags: [account]
      summary: 更新用户资料（管理端）
      x-function-id: F-1011
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                display_name:
                  type: string
                email:
                  type: string
                group:
                  type: string
                quota:
                  type: integer
                remark:
                  type: string
                status:
                  type: integer
              required: [id]
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /api/user/{id}:
    get:
      tags: [account]
      summary: 用户详情
      x-function-id: F-1008
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 用户详情（AdminView，含 setting）
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserAdminView'
  /api/user/manage:
    post:
      tags: [account]
      summary: 管理用户状态
      x-function-id: F-1010
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                action:
                  type: string
                  enum: [disable, enable, promote, demote, delete]
              required: [id, action]
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/user/{id}/credit:
    post:
      tags: [account]
      summary: 管理员给用户充值
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminBalanceAdjustRequest'
      responses:
        '200':
          description: 充值成功，data=充值后余额(USD)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: number
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/user/{id}/debit:
    post:
      tags: [account]
      summary: 管理员给用户扣费（扣到 0 为止）
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminBalanceAdjustRequest'
      responses:
        '200':
          description: 扣费成功，data=扣费后余额(USD)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: number
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/user/{id}/balance-logs:
    get:
      tags: [account]
      summary: 用户账变流水（充值/扣费/兑换/自助）
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: 账变流水列表（时间倒序）
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/BalanceTransactionView'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...